Extract a common domain primitives crate from system-runtime-primitives
No logical changes.

With the upcoming increasing domain runtime, this common domain primitives
crate will be used by all the domain runtimes. Mostly, it's enough to
just use this crate when creating a new domain runtime, unless this
domain runtime has some domain specific primitives, such as the system
domain's runtime api `SystemDomainApi`.
