+++ domains/client/domain-executor/src/system_domain_worker.rs
@@ -19,6 +19,7 @@ use crate::system_bundle_producer::SystemBundleProducer;
 use crate::utils::{BlockInfo, ExecutorSlotInfo};
 use crate::TransactionFor;
 use codec::{Decode, Encode};
+use domain_runtime_primitives::AccountId;
 use futures::channel::mpsc;
 use futures::{future, FutureExt, SinkExt, Stream, StreamExt, TryFutureExt};
 use sc_client_api::{AuxStore, BlockBackend, ProofProvider};
@@ -38,7 +39,7 @@ use std::future::Future;
 use std::pin::Pin;
 use std::sync::Arc;
 use subspace_core_primitives::{Blake2b256Hash, BlockNumber, Randomness};
-use system_runtime_primitives::{AccountId, SystemDomainApi};
+use system_runtime_primitives::SystemDomainApi;
 use tracing::Instrument;
 
 const LOG_TARGET: &str = "executor-worker";
