+++ domains/test/runtime/src/runtime.rs
@@ -1,3 +1,6 @@
+pub use domain_runtime_primitives::{
+    AccountId, Address, Balance, BlockNumber, Hash, Index, Signature,
+};
 use frame_support::traits::{ConstU16, ConstU32, Everything};
 use frame_support::weights::constants::{
     BlockExecutionWeight, ExtrinsicBaseWeight, RocksDbWeight, WEIGHT_PER_SECOND,
@@ -19,9 +22,6 @@ use sp_std::prelude::*;
 use sp_version::NativeVersion;
 use sp_version::RuntimeVersion;
 use subspace_runtime_primitives::{SHANNON, SSC};
-pub use system_runtime_primitives::{
-    AccountId, Address, Balance, BlockNumber, Hash, Index, Signature,
-};
 
 #[cfg(any(feature = "std", test))]
 pub use sp_runtime::BuildStorage;
@@ -414,7 +414,7 @@ impl_runtime_apis! {
         fn extract_signer(
             extrinsics: Vec<<Block as BlockT>::Extrinsic>,
         ) -> Vec<(Option<AccountId>, <Block as BlockT>::Extrinsic)> {
-            use system_runtime_primitives::Signer;
+            use domain_runtime_primitives::Signer;
             let lookup = frame_system::ChainContext::<Runtime>::default();
             extrinsics.into_iter().map(|xt| (xt.signer(&lookup), xt)).collect()
         }
