+++ test/subspace-test-runtime/Cargo.toml
@@ -17,6 +17,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
 codec = { package = "parity-scale-codec", version = "3.1.5", default-features = false, features = ["derive"] }
+domain-runtime-primitives = { version = "0.1.0", default-features = false, path = "../../domains/primitives/runtime" }
 domain-test-runtime = { version = "0.1.0", default-features = false, path = "../../domains/test/runtime" }
 frame-executive = { version = "4.0.0-dev", default-features = false, git = "https://github.com/subspace/substrate", rev = "1a7c28721fa77ecce9632ad9ce473f2d3cf1a598" }
 frame-support = { version = "4.0.0-dev", default-features = false, git = "https://github.com/subspace/substrate", rev = "1a7c28721fa77ecce9632ad9ce473f2d3cf1a598" }
@@ -53,7 +54,6 @@ sp-version = { version = "5.0.0", default-features = false, git = "https://githu
 subspace-core-primitives = { version = "0.1.0", default-features = false, path = "../../crates/subspace-core-primitives" }
 subspace-runtime-primitives = { version = "0.1.0", default-features = false, path = "../../crates/subspace-runtime-primitives" }
 subspace-verification = { version = "0.1.0", default-features = false, path = "../../crates/subspace-verification" }
-system-runtime-primitives = { version = "0.1.0", default-features = false, path = "../../domains/primitives/system-runtime" }
 
 # Used for the node template's RPCs
 frame-system-rpc-runtime-api = { version = "4.0.0-dev", default-features = false, git = "https://github.com/subspace/substrate", rev = "1a7c28721fa77ecce9632ad9ce473f2d3cf1a598" }
@@ -66,8 +66,9 @@ substrate-wasm-builder = { version = "5.0.0-dev", git = "https://github.com/subs
 [features]
 default = ["std"]
 std = [
-	"domain-test-runtime/std",
 	"codec/std",
+	"domain-runtime-primitives/std",
+	"domain-test-runtime/std",
 	"frame-executive/std",
 	"frame-support/std",
 	"frame-system/std",
