+++ test/subspace-test-runtime/src/lib.rs
@@ -474,7 +474,7 @@ parameter_types! {
 
 impl pallet_domains::Config for Runtime {
     type Event = Event;
-    type SecondaryHash = system_runtime_primitives::Hash;
+    type SecondaryHash = domain_runtime_primitives::Hash;
     type ReceiptsPruningDepth = ReceiptsPruningDepth;
     type MaximumReceiptDrift = MaximumReceiptDrift;
     type ConfirmationDepthK = ConfirmationDepthK;
@@ -817,8 +817,8 @@ fn extract_block_object_mapping(block: Block, successful_calls: Vec<Hash>) -> Bl
 fn extract_system_bundles(
     extrinsics: Vec<UncheckedExtrinsic>,
 ) -> (
-    sp_domains::OpaqueBundles<Block, system_runtime_primitives::Hash>,
-    sp_domains::SignedOpaqueBundles<Block, system_runtime_primitives::Hash>,
+    sp_domains::OpaqueBundles<Block, domain_runtime_primitives::Hash>,
+    sp_domains::SignedOpaqueBundles<Block, domain_runtime_primitives::Hash>,
 ) {
     let (system_bundles, core_bundles): (Vec<_>, Vec<_>) = extrinsics
         .into_iter()
@@ -846,7 +846,7 @@ fn extract_system_bundles(
 fn extract_core_bundles(
     extrinsics: Vec<UncheckedExtrinsic>,
     domain_id: DomainId,
-) -> sp_domains::OpaqueBundles<Block, system_runtime_primitives::Hash> {
+) -> sp_domains::OpaqueBundles<Block, domain_runtime_primitives::Hash> {
     extrinsics
         .into_iter()
         .filter_map(|uxt| match uxt.function {
@@ -862,7 +862,7 @@ fn extract_core_bundles(
 
 fn extract_receipts(
     extrinsics: Vec<UncheckedExtrinsic>,
-) -> Vec<ExecutionReceipt<BlockNumber, Hash, system_runtime_primitives::Hash>> {
+) -> Vec<ExecutionReceipt<BlockNumber, Hash, domain_runtime_primitives::Hash>> {
     extrinsics
         .into_iter()
         .filter_map(|uxt| {
@@ -1086,8 +1086,8 @@ impl_runtime_apis! {
         }
     }
 
-    impl sp_domains::ExecutorApi<Block, system_runtime_primitives::Hash> for Runtime {
-        fn submit_bundle_unsigned(opaque_bundle: SignedOpaqueBundle<NumberFor<Block>, <Block as BlockT>::Hash, system_runtime_primitives::Hash>) {
+    impl sp_domains::ExecutorApi<Block, domain_runtime_primitives::Hash> for Runtime {
+        fn submit_bundle_unsigned(opaque_bundle: SignedOpaqueBundle<NumberFor<Block>, <Block as BlockT>::Hash, domain_runtime_primitives::Hash>) {
             Domains::submit_bundle_unsigned(opaque_bundle)
         }
 
@@ -1110,8 +1110,8 @@ impl_runtime_apis! {
         fn extract_system_bundles(
             extrinsics: Vec<<Block as BlockT>::Extrinsic>,
         ) -> (
-            sp_domains::OpaqueBundles<Block, system_runtime_primitives::Hash>,
-            sp_domains::SignedOpaqueBundles<Block, system_runtime_primitives::Hash>,
+            sp_domains::OpaqueBundles<Block, domain_runtime_primitives::Hash>,
+            sp_domains::SignedOpaqueBundles<Block, domain_runtime_primitives::Hash>,
         ) {
             extract_system_bundles(extrinsics)
         }
@@ -1119,13 +1119,13 @@ impl_runtime_apis! {
         fn extract_core_bundles(
             extrinsics: Vec<<Block as BlockT>::Extrinsic>,
             domain_id: DomainId,
-        ) -> sp_domains::OpaqueBundles<Block, system_runtime_primitives::Hash> {
+        ) -> sp_domains::OpaqueBundles<Block, domain_runtime_primitives::Hash> {
             extract_core_bundles(extrinsics, domain_id)
         }
 
         fn extract_receipts(
             extrinsics: Vec<<Block as BlockT>::Extrinsic>,
-        ) -> Vec<ExecutionReceipt<NumberFor<Block>, <Block as BlockT>::Hash, system_runtime_primitives::Hash>> {
+        ) -> Vec<ExecutionReceipt<NumberFor<Block>, <Block as BlockT>::Hash, domain_runtime_primitives::Hash>> {
             extract_receipts(extrinsics)
         }
